+++ src/main/php/org/nagios/nsca/NscaMessage.class.php
@@ -35,6 +35,7 @@ class NscaMessage extends \lang\Object {
   /**
    * Retrieve a status' name
    *
+   * @deprecated Use NscaProtocol::statusName()
    * @param   int status
    * @return  string name
    */
@@ -87,7 +88,7 @@ class NscaMessage extends \lang\Object {
   /**
    * Set Status
    *
-   * @param   int status one of NSCA_* constants
+   * @param   int status one of NscaProtocol's class constants OK, WARN, ERROR and UNKNOWN
    */
   public function setStatus($status) {
     $this->status= $status;
@@ -96,7 +97,7 @@ class NscaMessage extends \lang\Object {
   /**
    * Get Status
    *
-   * @return  int one of NSCA_* constants
+   * @return  int
    */
   public function getStatus() {
     return $this->status;